hhkb-studio: update livecheck
The existing `livecheck` block for `hhkb-studio` is currently giving
an `Unable to get versions` error, as the regex isn't matching the
current version text.

This updates the regex to match the HHKB Studio Keymap Tool version
for macOS. This works for now but it's a loose regex that can
potentially lead to unexpected matches if the page structure changes
in the future. Unfortunately, this may be the best we can do for the
moment (as the filename doesn't use dots in the version).